+++ R/admiral-package.R
@@ -5,13 +5,13 @@
 #'             summarise_at summarise coalesce bind_cols na_if tibble
 #' @importFrom magrittr %>%
 #' @importFrom rlang := abort arg_match as_function as_string call2 caller_env
-#'             call_name .data enexpr enquo eval_bare eval_tidy expr expr_interp
-#'             expr_label exprs f_lhs f_rhs friendly_type inform is_bare_formula
-#'             is_call is_character is_formula is_integerish is_quosure
-#'             is_quosures is_symbol new_formula parse_exprs quo quo_get_expr
-#'             quo_is_call quo_is_missing quo_is_null quo_is_symbol quos
-#'             quo_squash quo_text set_names sym syms type_of warn quo_set_env
-#'             quo_get_env
+#'             call_name current_env .data enexpr enquo eval_bare eval_tidy expr
+#'             expr_interp expr_label exprs f_lhs f_rhs friendly_type inform
+#'             is_bare_formula is_call is_character is_formula is_integerish
+#'             is_quosure is_quosures is_symbol new_formula parse_exprs quo
+#'             quo_get_expr quo_is_call quo_is_missing quo_is_null quo_is_symbol
+#'             quos quo_squash quo_text set_names sym syms type_of warn
+#'             quo_set_env quo_get_env
 #' @importFrom utils capture.output
 #' @importFrom purrr map map2 map_chr map_lgl reduce walk keep map_if transpose
 #'             flatten every modify_at modify_if reduce compose